writing-practice/requirements.txt
streamlit==1.22.0
requests==2.28.2
pyyaml==6.0
manga-ocr==0.1.8
Pillow==9.4.0
flask==2.0.0
opencv-python==4.7.0
streamlit-drawable-canvas==0.9.0
# ollama>=0.1.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1
